Create nn_attempt.py
Neural networks seem to perform particularly poorly at separating the data set. The SVM with kernel trick is able to elevate the problem to a high enough order dimensions that it becomes linearly separable in a meaningful way. Attempts to do the same by adding increasing number of hidden layers does not seem to work nearly as well, because the NNs constantly get trapped at poor performing local minima.